+++ catalog/src/main/java/it/eng/catalog/service/CatalogService.java
@@ -226,32 +226,31 @@ public class CatalogService {
      * @param offer
      * @return boolean
      */
-	public boolean validateOffer(Offer offer) {
-		boolean valid = false;
+    public boolean validateOffer(Offer offer) {
+    	boolean valid = false;
 		Catalog catalog = getCatalog();
 		Dataset dataset = catalog.getDataset().stream()
 				.filter(ds -> ds.getId().equals(offer.getTarget())).findFirst()
 				.orElse(null);
-		if(dataset == null) {
+		if (dataset == null) {
 			log.warn("Offer.target '{}' does not match with any dataset from catalog", offer.getTarget());
-			return false;
-		}
-
-		Offer existingOffer = dataset.getHasPolicy().stream()
-				.filter(of -> of.getId().equals(offer.getId()))
-				.findFirst()
-				.orElse(null);
-
-		log.debug("Offer with id '{}' {}", offer.getId(), existingOffer != null ? " found." : "not found.");
-		
-		if (existingOffer == null) {
-			log.warn("Offer with id {} not found in catalog", offer.getId());
-			valid = false;
 		} else {
-//	    		 check if offers are equals
-			if (offer.equals(existingOffer)) {
-				log.debug("Existing and prvided offers are same");
-				valid = true;
+
+			Offer existingOffer = dataset.getHasPolicy().stream()
+					.filter(of -> of.getId().equals(offer.getId()))
+					.findFirst()
+					.orElse(null);
+
+			log.debug("Offer with id '{}' {}", offer.getId(), existingOffer != null ? " found." : "not found.");
+
+			if (existingOffer == null) {
+				log.warn("Offer with id {} not found in catalog", offer.getId());
+			} else {
+				// check if offers are equals
+				if (offer.equals(existingOffer)) {
+					log.debug("Existing and provided offers are same");
+					valid = true;
+				}
 			}
 		}
 		log.info("Offer evaluated as {}", valid ? "valid" : "invalid");